add gen_subaddress to identifier, re-structure identifier module
